+++ crates/zng-app/src/running.rs
@@ -1186,7 +1186,7 @@ command! {
     pub static EXIT_CMD = {
         l10n!: true,
         name: "Exit",
-        info: "Close all windows and exit.",
+        info: "Close all windows and exit",
         shortcut: shortcut!(Exit),
     };
 }