+++ crates/zng-ext-clipboard/src/lib.rs
@@ -428,7 +428,7 @@ command! {
     pub static CUT_CMD = {
         l10n!: true,
         name: "Cut",
-        info: "Remove the selection and place it in the clipboard.",
+        info: "Remove the selection and place it in the clipboard",
         shortcut: [shortcut!(CTRL+'X'), shortcut!(SHIFT+Delete), shortcut!(Cut)],
         shortcut_filter: ShortcutFilter::FOCUSED | ShortcutFilter::CMD_ENABLED,
         icon: wgt_fn!(|_| ICONS.get("cut")),
@@ -438,7 +438,7 @@ command! {
     pub static COPY_CMD = {
         l10n!: true,
         name: "Copy",
-        info: "Place a copy of the selection in the clipboard.",
+        info: "Place a copy of the selection in the clipboard",
         shortcut: [shortcut!(CTRL+'C'), shortcut!(CTRL+Insert), shortcut!(Copy)],
         shortcut_filter: ShortcutFilter::FOCUSED | ShortcutFilter::CMD_ENABLED,
         icon: wgt_fn!(|_| ICONS.get("copy")),
@@ -448,7 +448,7 @@ command! {
     pub static PASTE_CMD = {
         l10n!: true,
         name: "Paste",
-        info: "Insert content from the clipboard.",
+        info: "Insert content from the clipboard",
         shortcut: [shortcut!(CTRL+'V'), shortcut!(SHIFT+Insert), shortcut!(Paste)],
         shortcut_filter: ShortcutFilter::FOCUSED | ShortcutFilter::CMD_ENABLED,
         icon: wgt_fn!(|_| ICONS.get("paste")),