+++ crates/zng-ext-window/src/cmd.rs
@@ -19,7 +19,7 @@ command! {
     pub static CLOSE_CMD = {
         l10n!: true,
         name: "Close",
-        info: "Close the window.",
+        info: "Close the window",
         shortcut: [shortcut!(ALT+F4), shortcut!(CTRL+'W')],
         icon: wgt_fn!(|_| ICONS.get(["window-close", "close"])),
     };
@@ -28,7 +28,7 @@ command! {
     pub static MINIMIZE_CMD = {
         l10n!: true,
         name: "Minimize",
-        info: "Minimize the window.",
+        info: "Minimize the window",
         icon: wgt_fn!(|_| ICONS.get(["window-minimize"])),
     };
 
@@ -36,7 +36,7 @@ command! {
     pub static MAXIMIZE_CMD = {
         l10n!: true,
         name: "Maximize",
-        info: "Maximize the window.",
+        info: "Maximize the window",
         icon: wgt_fn!(|_| ICONS.get(["window-maximize"])),
     };
 
@@ -49,7 +49,7 @@ command! {
     pub static FULLSCREEN_CMD = {
         l10n!: true,
         name: "Fullscreen",
-        info: "Toggle fullscreen mode on the window.",
+        info: "Toggle fullscreen mode on the window",
         shortcut: {
             let a = if cfg!(target_os = "macos") {
                 shortcut!(CTRL|SHIFT+'F')
@@ -70,7 +70,7 @@ command! {
     pub static EXCLUSIVE_FULLSCREEN_CMD = {
         l10n!: true,
         name: "Exclusive Fullscreen",
-        info: "Toggle exclusive fullscreen mode on the window.",
+        info: "Toggle exclusive fullscreen mode on the window",
         icon: wgt_fn!(|_| ICONS.get(["window-exclusive-fullscreen", "window-fullscreen", "fullscreen"])),
     };
 
@@ -80,7 +80,7 @@ command! {
     pub static RESTORE_CMD = {
         l10n!: true,
         name: "Restore",
-        info: "Restores the window to its previous non-minimized state or normal state.",
+        info: "Restores the window to its previous non-minimized state or normal state",
         icon: wgt_fn!(|_| ICONS.get(["window-restore"])),
     };