+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+
+# 0.11.7
+
 * Fix OpenGL version check.
 * Fix window receiving a cursor move event while cursor is not over (on x11).
 * Remove end punctuation from command `info` fields